requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic[email]==2.5.3
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3

# Environment variables
python-dotenv==1.0.0
